Small fixes to simngs and fames tests

+++ tests/test_simngs.py
@@ -36,9 +36,10 @@ class SimNGSTest(unittest.TestCase):
             os.chdir("src")
             subprocess.check_call(["make"])
 
-        shutil.move(simngs, self.progs)
-        shutil.move(simlib, self.progs)
-        shutil.move(runfile, self.progs)
+        if not os.path.exists(dirname):
+            shutil.move(simngs, self.progs)
+            shutil.move(simlib, self.progs)
+            shutil.move(runfile, self.progs)
 
     def test_2_run_simNGS(self):
         """ Generates a synthetic library and runs with built-in simNGS runfile
